test: move pig latin tests onto shared fixture
