+++ .github/workflows/e2e-piglatin.yml
@@ -44,7 +44,7 @@ jobs:
 
       - name: Run pig-latin SFT end-to-end
         working-directory: examples
-        run: uv run python -m unittest discover -s sft/pig-latin/tests -v
+        run: PYTHONPATH=.:sft/pig-latin uv run python -m unittest tests.test_piglatin_qwen -v
         env:
           TINKER_API_KEY: tml-dummy-key
           TRANSFORMERS_VERBOSITY: error
