+++ examples/tests/test_piglatin_gemma.py
@@ -6,9 +6,15 @@ preset's threshold, and the final exact-match clears a 20% floor.
 """
 
 import unittest
+from pathlib import Path
 
-from _server_fixture import CLIENT_DIR, PIGLATIN_EVAL_EXAMPLES, OpenRlServerCase
 from piglatin_sft import PRESETS, run_training
+from test_piglatin_qwen import PIGLATIN_EVAL_EXAMPLES
+
+from tests._server_fixture import OpenRlServerCase
+
+REPO_ROOT = Path(__file__).resolve().parents[2]
+CLIENT_DIR = REPO_ROOT / "examples" / "sft" / "pig-latin"
 
 
 class TestPigLatinGemma(OpenRlServerCase):
@@ -17,27 +23,23 @@ class TestPigLatinGemma(OpenRlServerCase):
   REQUIRE_HF_TOKEN = True
 
   def test_sft_improves(self) -> None:
-    config = (
-      PRESETS["gemma"]
-      .clone()
-      .apply(
-        {
-          "base_url": self.BASE_URL,
-          "steps": 25,
-          "assert_improvement": False,
-          "plot_path": str(CLIENT_DIR / "artifacts" / "test_gemma_metrics.png"),
-          "custom_examples": PIGLATIN_EVAL_EXAMPLES,
-        },
-        layer_name="test_override",
-      )
-      .make()
-    )
+    TEST_CONFIG = {
+      "base_url": self.BASE_URL,
+      "steps": 25,
+      "assert_improvement": False,
+      "plot_path": str(CLIENT_DIR / "artifacts" / "test_gemma_metrics.png"),
+      "custom_examples": PIGLATIN_EVAL_EXAMPLES,
+    }
+
+    config = PRESETS["gemma"].clone().apply(TEST_CONFIG, layer_name="test_override").make()
 
     m = run_training(config)
     print(f"Training metrics: {m}")
 
     self.assertGreater(m["after_exact"], m["before_exact"], f"Exact match didn't improve: {m['before_exact']:.2f} -> {m['after_exact']:.2f}")
+
     sim_gain = m["after_sim"] - m["before_sim"]
+
     self.assertGreaterEqual(sim_gain, config.min_similarity_gain, f"Similarity gain insufficient: {m['before_sim']:.2f} -> {m['after_sim']:.2f}")
     self.assertGreaterEqual(m["after_exact"], 0.20, f"Expected >= 20% exact match, got {m['after_exact']:.0%}")
